Added a way to view the new user ratings in videoinfo
http://forum.kodi.tv/showthread.php?tid=250862&pid=2179671#pid2179671

Basically, if there is no "next aired", then show the user rating, but
if there is next aired, show user rating when user rating button is
focused.  Same with movie info in that if no rating is available then
show always, otherwise only show when user rating button is focused.